\name{SeqAlloc-package}
\alias{SeqAlloc-package}
\docType{package}
\title{
Sequential Allocation for Randomized Experiments
}
\description{
Potential randomization schemes are prospectively evaluated when units are assigned to treatment arms upon entry into the experiment. The schemes are evaluated for balance on covariates and on predictability (i.e., how well could a site worker guess the treatment of the next unit enrolled).
}
\details{
\tabular{ll}{
Package: \tab SeqAlloc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-08-03\cr
Imports: \tab stats, graphics\cr
License: \tab GPL-2\cr
}
}
\author{
Xiaoshu Zhu \email{xiaoshuzhu@westat.com} and Sharon Lohr
}
\references{
Lohr, S. and X. Zhu (2015). Randomized Sequential Individual Assignment in Social Experiments: Evaluating the Design Options Prospectively. Sociological Methods and Research. [Advance online publication: December 27, 2015] doi: 10.1177/0049124115621332.

McEntegart, D. J. (2003). The Pursuit of Balance Using Stratified and Dynamic Randomization Techniques: An Overview. Drug information Journal 37: 293-308.

Rosenberger, W. F. and J. M. Lachin (2004). Randomization in Clinical Trials: Theory and Practice. New York: Wiley.
}